% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_util.R
\name{AssociatedDimReducs}
\alias{AssociatedDimReducs}
\title{Get dimensional reduction information associated with an assay}
\usage{
AssociatedDimReducs(
  object,
  assay = DefaultAssay(object = object),
  global = TRUE
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{Name of assay that dimensional reduction objects should be
associated with}

\item{global}{Include global dimensional reductions}
}
\value{
A vector of dimensional reduction names
}
\description{
Get dimensional reduction information associated with an assay
}
\references{
https://github.com/satijalab/seurat-wrappers/wiki/Code-Guidelines
}
\author{
Paul Hoffman
}
\keyword{internal}
