% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_service.R
\name{atac_qc_list}
\alias{atac_qc_list}
\title{Get QC metrics list for single scRNA-seq dataset}
\usage{
atac_qc_list()
}
\value{
list of QC metrics for plotting
}
\description{
Get QC metrics list for single scRNA-seq dataset
}
