% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{BinMean}
\alias{BinMean}
\title{Calculate mean values of a matrix by bins}
\usage{
BinMean(vec, every, na.rm = FALSE)
}
\arguments{
\item{vec}{vector}

\item{every}{number bin size}
}
\value{

}
\description{
Calculate mean values of a matrix by bins
}
