% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_service.R
\name{calc_deg}
\alias{calc_deg}
\title{Run Seurat clustering}
\usage{
calc_deg(
  req,
  ident1_idx = 5,
  ident2_idx = 5,
  ident1 = 4,
  ident2 = 5,
  min_pct = 0.2,
  min_lfc = 0.5
)
}
\arguments{
\item{req}{request payload}

\item{ident1_idx}{string}

\item{ident2_idx}{string}

\item{ident1}{string}

\item{ident2}{string}

\item{min_pct}{string}

\item{min_lfc}{string}
}
\value{

}
\description{
Run Seurat clustering
}
