% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_service.R
\name{calc_dr}
\alias{calc_dr}
\title{Run Seurat clustering}
\usage{
calc_dr(req, ident1 = 4, ident2 = 5)
}
\arguments{
\item{req}{request payload}

\item{ident1}{string}

\item{ident2}{string}

\item{min_pct}{string}

\item{min_lfc}{string}

\item{assay}{string}

\item{pvalue}{string}
}
\value{

}
\description{
Run Seurat clustering
}
