% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{calc_enrichr_table}
\alias{calc_enrichr_table}
\title{Run Enrichr}
\usage{
calc_enrichr_table(req, genes = c("CD74", "CD7"), database = "KEGG")
}
\arguments{
\item{req}{request payload}

\item{genes}{array}

\item{database}{string}
}
\value{

}
\description{
Run Enrichr
}
