% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{calc_gsea_plot}
\alias{calc_gsea_plot}
\title{Generate GSEA plot}
\usage{
calc_gsea_plot(req, term)
}
\arguments{
\item{req}{request payload}

\item{genes}{string}

\item{database}{string}
}
\value{

}
\description{
Generate GSEA plot
}
