% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{calc_gsea_table}
\alias{calc_gsea_table}
\title{Run GSEA enrichment}
\usage{
calc_gsea_table(req, genes, database)
}
\arguments{
\item{req}{request payload}

\item{genes}{string}

\item{database}{string}
}
\value{

}
\description{
Run GSEA enrichment
}
