% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_package}
\alias{check_package}
\title{Check to ensure a package is installed}
\usage{
check_package(package, repository, ...)
}
\arguments{
\item{package}{Name of pacakge to check}

\item{repository}{Repository that package is available on;
choose from 'bioconductor', 'github', or 'cran'}

\item{...}{Extra parameters passed to BiocManager::install,
remotes::install_github, or install.packages, depending on \code{repository}}
}
\description{
Check to ensure a package is installed
}
\references{
https://github.com/satijalab/seurat-wrappers/wiki/Code-Guidelines
}
\author{
Paul Hoffman
}
