% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{cluster_coords}
\alias{cluster_coords}
\title{Return embeddings coordinates from a given embedding name}
\usage{
cluster_coords(req, categoryName = "Sex", embedding = "umap")
}
\arguments{
\item{req}{request payload}

\item{categoryName}{string}

\item{embedding}{string}
}
\value{
json
}
\description{
Return embeddings coordinates from a given embedding name
}
