% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Run Seurat clustering}
\usage{
cluster_seurat(req, nPCs = 20, resolution = 0.5)
}
\arguments{
\item{req}{test}

\item{nPCs}{test}

\item{resolution}{test}
}
\value{

}
\description{
Run Seurat clustering
}
