% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{cluster_single_rna}
\alias{cluster_single_rna}
\title{Run Seurat clustering}
\usage{
cluster_single_rna(
  req,
  jobid = "example",
  load = "load",
  nPCs = 15,
  resolution = 0.5,
  neighbor = 20,
  expr = NULL
)
}
\arguments{
\item{req}{request payload}

\item{jobid}{string}

\item{nPCs}{string}

\item{resolution}{string}

\item{neighbor}{string}

\item{expr}{}
}
\value{

}
\description{
Run Seurat clustering
}
