% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_empty}
\alias{convert_empty}
\title{Convert empty string to "-"}
\usage{
convert_empty(str)
}
\arguments{
\item{str}{string}
}
\value{

}
\description{
Convert empty string to "-"
}
\references{
https://rviews.rstudio.com/2019/08/13/plumber-logging/
}
