% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{coverage_plot}
\alias{coverage_plot}
\title{ATAC coverage static plot}
\usage{
coverage_plot(
  req,
  type = "gene",
  gene = "GAD2",
  flank = "10000",
  chr = "chr10",
  start = "26116307",
  end = "26316307",
  is_annotation = T,
  is_peak = F
)
}
\arguments{
\item{req}{request payload}

\item{type}{string}

\item{gene}{string#'}

\item{flank}{string}

\item{chr}{string}

\item{start}{string}

\item{end}{string}

\item{is_annotation}{string}

\item{is_peak}{string}
}
\value{
static plot
}
\description{
ATAC coverage static plot
}
