% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_uploadinfo_object}
\alias{create_uploadinfo_object}
\title{create_uploadinfo_object}
\usage{
create_uploadinfo_object(upload_info = list())
}
\arguments{
\item{upload_info}{upload data information}
}
\value{
a S4 object.
}
\description{
Create IRIS3 UploadInfo object
}
\details{
Create UploadInfo Object
}
