% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateIRIS3Object}
\alias{CreateIRIS3Object}
\title{CreateIRIS3Object}
\usage{
CreateIRIS3Object(x = seurat_object, upload_info = new(Class = "UploadInfo"))
}
\arguments{
\item{x}{Seurat, a seurat object}

\item{upload_info}{upload data information}
}
\value{
it should return a IRIS3 S4 object.
}
\description{
Create IRIS3 object
}
\details{
CreateIRIS3Object
}
\examples{
data("yan_2013")
seurat_obj <- Seurat::CreateSeuratObject(yan_2013$expr)
object <- CreateIRIS3Object(seurat_obj)
}
