% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_util.R
\name{DefaultDimReduc}
\alias{DefaultDimReduc}
\title{Find the default DimReduc}
\usage{
DefaultDimReduc(object, assay = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use;
defaults to the default assay of the object}
}
\value{
The default DimReduc, if possible
}
\description{
Searches for DimReducs matching 'umap', 'tsne', or 'pca',
case-insensitive, and in that order. Priority given to
DimReducs matching the DefaultAssay or assay specified
(eg. 'pca' for the default assay weights higher than
'umap' for a non-default assay)
}
\references{
https://github.com/satijalab/seurat-wrappers/wiki/Code-Guidelines
}
\author{
Paul Hoffman
}
