% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_service.R
\name{example_dr_figure}
\alias{example_dr_figure}
\title{Run DR peak-gene figure}
\usage{
example_dr_figure(
  genes = c("CTCF", "ELF1", "MEF2C", "E2F6", "EGR1"),
  ct1 = c(4),
  ct2 = c(1)
)
}
\value{

}
\description{
Run DR peak-gene figure
}
