% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_service.R
\name{example_ri_heatmap2}
\alias{example_ri_heatmap2}
\title{Run RI heatmap}
\usage{
example_ri_heatmap2(
  tf = "CTCF",
  genes = c("Arhgap27", "Zbtb22", "Fam126a", "Mfsd14b")
)
}
\value{

}
\description{
Run RI heatmap
}
