% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{feature_coords}
\alias{feature_coords}
\title{Return embeddings coordinates and feature expression}
\usage{
feature_coords(req, gene = "Gad1", assay = "RNA")
}
\arguments{
\item{req}{request payload}

\item{categoryName}{string}

\item{embedding}{string}
}
\value{
json
}
\description{
Return embeddings coordinates and feature expression
}
