% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Complex heatmap}
\usage{
feature_heatmap(
  req,
  features = c("Gad1", "Gad2"),
  meta = "cell_type",
  color = NULL
)
}
\arguments{
\item{req}{request payload}

\item{features}{vector of string}

\item{meta}{metadata to annotate in columns}

\item{color}{color palette}
}
\value{
static image
}
\description{
Complex heatmap
}
