% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{gene_cor_plot}
\alias{gene_cor_plot}
\title{Gene-gene correlation static plot}
\usage{
gene_cor_plot(req, gene1 = "Gad1", gene2 = "Gad2")
}
\arguments{
\item{req}{request payload}

\item{gene1}{string gene1}

\item{gene2}{string gene2}
}
\value{
static plot
}
\description{
Gene-gene correlation static plot
}
