% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{gene_umap_plot}
\alias{gene_umap_plot}
\title{Plot static UMAP}
\usage{
gene_umap_plot(req, gene = "Gad1", assay = "RNA")
}
\arguments{
\item{req}{request payload}

\item{gene}{string}

\item{assay}{string}
}
\value{
static image
}
\description{
Plot static UMAP
}
