% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_service.R
\name{get_all_embeddings}
\alias{get_all_embeddings}
\title{Get all embedding/dimension reduction names from object}
\usage{
get_all_embeddings()
}
\value{
array all embedding names
}
\description{
Get all embedding/dimension reduction names from object
}
