% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_gene_range}
\alias{get_gene_range}
\title{Get genomic range of a gene and extend a flank region}
\usage{
get_gene_range(gene = "GAD1", flank = "10000")
}
\arguments{
\item{gene}{string}

\item{flank}{string}
}
\value{

}
\description{
Get genomic range of a gene and extend a flank region
}
