% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getdata}
\alias{getdata}
\title{Load dataset from "R" package using data(), assign it directly to a variable}
\usage{
getdata(...)
}
\value{

}
\description{
Load dataset from "R" package using data(), assign it directly to a variable
}
