% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hello_name}
\alias{hello_name}
\title{Hello world function}
\usage{
hello_name(name)
}
\arguments{
\item{name}{echo name}
}
\value{
A string
}
\description{
This function sends hello message from a given name.
}
