% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifnb_2800}
\alias{ifnb_2800}
\title{Data of ifnb_2800}
\format{
A list with expression matrix and metadata:
#'\describe{
\item{$expr}{a data frame with x genes (rows) amd 2800 cells (columns)}
\item{$meta$Label}{fct cell type}
\item{$meta$Sex}{fct random generated sex info}
\item{$meta$Sample}{fct random generated sample info}
}
}
\source{
\url{https://satijalab.org/seurat/articles/integration_introduction.html}
}
\usage{
ifnb_2800
}
\description{
ifnb 2800: Subset of IFNB-Stimulated and Control PBMCs, Human, 15 cell types, 2800 cells
}
\keyword{datasets}
