% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{interactive_heatmap}
\alias{interactive_heatmap}
\title{Interactive complex heatmap}
\usage{
interactive_heatmap(
  req,
  genes = c("CD74", "CD7"),
  meta = "cell_type",
  color = NULL
)
}
\arguments{
\item{req}{request payload}
}
\value{
html
}
\description{
Interactive complex heatmap
}
