% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris3api-package.R
\docType{package}
\name{iris3api-package}
\alias{iris3api}
\alias{iris3api-package}
\title{iris3api: IRIS3-api}
\description{
The R API server for IRIS3 website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Wang-Cankun/iris3api}
  \item Report bugs at \url{https://github.com/Wang-Cankun/iris3api/issues}
}

}
\author{
\strong{Maintainer}: Cankun Wang \email{flykun0620@gmail.com} (\href{https://orcid.org/0000-0002-0225-9855}{ORCID})

}
\keyword{internal}
