% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_matrix_empty}
\alias{is_matrix_empty}
\title{Check if a matrix is empty}
\usage{
is_matrix_empty(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Whether or not \code{x} is empty
}
\description{
Takes a matrix and asks if it's empty (either 0x0 or 1x1 with a value of NA)
}
\references{
https://github.com/satijalab/seurat-wrappers/wiki/Code-Guidelines
}
\author{
Paul Hoffman
}
