% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_multi_rna}
\alias{load_multi_rna}
\title{Load multi scRNA-seq data to Seurat}
\usage{
load_multi_rna(
  req,
  idx = 1,
  jobid = "1642208223822",
  type = "multiRna",
  mode = "RNA",
  load = "load",
  filename,
  min_cells = 20000,
  min_genes = 0.001,
  nVariableFeatures = 2000,
  percentMt = 0.1,
  percentRb = 0.5,
  removeOutlier = FALSE,
  label = NULL,
  species = "Human"
)
}
\arguments{
\item{req}{request payload}

\item{idx}{sample index}

\item{filename}{string}

\item{min_cells}{number}

\item{min_genes}{number}

\item{nVariableFeatures}{number}

\item{percentMt}{number}

\item{removeRibosome}{boolean}
}
\value{
list of basic QC metrics
}
\description{
Load multi scRNA-seq data to Seurat
}
