% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_multiome}
\alias{load_multiome}
\title{Load matched scRNA-seq and scATAC-seq data to Seurat / Signac}
\usage{
load_multiome(
  req,
  idx = 0,
  filename,
  jobid = "example",
  mode = "RNA",
  min_cells = 0.001,
  min_counts = 20000,
  removeOutlier = T,
  nVariableFeatures = 2000,
  percentMt = 0.25,
  percentRb = 0.5,
  expr = NULL,
  label = NULL,
  species = "Human",
  destination = NULL
)
}
\arguments{
\item{req}{request payload}

\item{idx}{sample index}

\item{filename}{string}

\item{mode}{string}

\item{min_cells}{number}

\item{min_counts}{number}

\item{removeOutlier}{boolean}

\item{nVariableFeatures}{number}

\item{percentMt}{number}
}
\value{
list of basic QC metrics
}
\description{
Load matched scRNA-seq and scATAC-seq data to Seurat / Signac
}
