% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_multiome2}
\alias{load_multiome2}
\title{Load matched scRNA-seq and scATAC-seq data to Seurat / Signac}
\usage{
load_multiome2(
  req,
  idx = 0,
  filename,
  jobid = "example",
  mode = "ATAC",
  min_cells = 1,
  min_genes = 200,
  nVariableFeatures = 2000,
  percentMt = 25,
  removeRibosome = FALSE,
  expr = NULL,
  label = NULL,
  species = "Human"
)
}
\arguments{
\item{req}{request payload}

\item{idx}{sample index}

\item{filename}{string}

\item{mode}{string}

\item{min_cells}{number}

\item{min_genes}{number}

\item{nVariableFeatures}{number}

\item{percentMt}{number}

\item{removeRibosome}{boolean}
}
\value{
list of basic QC metrics
}
\description{
Load matched scRNA-seq and scATAC-seq data to Seurat / Signac
}
