% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_seurat}
\alias{load_seurat}
\title{Load data to Seurat}
\usage{
load_seurat(
  req,
  filename,
  min_cells = 1,
  min_genes = 200,
  nVariableFeatures = 3000,
  percentMt = 5,
  removeRibosome = FALSE
)
}
\arguments{
\item{req}{test}

\item{filename}{test}

\item{min_cells}{test}

\item{min_genes}{test}

\item{nVariableFeatures}{test}

\item{percentMt}{test}

\item{removeRibosome}{test}
}
\value{
list of basic QC metrics
}
\description{
Load data to Seurat
}
