% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_single_rna}
\alias{load_single_rna}
\title{Load data to Seurat}
\usage{
load_single_rna(
  req,
  idx = 1,
  jobid = "example",
  type = "eg",
  min_cells = 1,
  min_genes = 200,
  nVariableFeatures = 2000,
  percentMt = 5,
  removeRibosome = FALSE,
  expr = NULL,
  label = NULL,
  species = "Human"
)
}
\arguments{
\item{req}{request payload}

\item{idx}{sample index}

\item{min_cells}{number}

\item{min_genes}{number}

\item{nVariableFeatures}{number}

\item{percentMt}{number}

\item{removeRibosome}{boolean}

\item{filename}{string}
}
\value{
list of basic QC metrics
}
\description{
Load data to Seurat
}
