% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{merge_idents}
\alias{merge_idents}
\title{Merge active clusters}
\usage{
merge_idents(req, newClusterIds)
}
\arguments{
\item{req}{request payload}

\item{newClusterIds}{array}
}
\value{
array levels of renamed new idents
}
\description{
Merge active clusters
}
