% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_service.R
\name{plot_cellchat}
\alias{plot_cellchat}
\title{Run CellChat}
\usage{
plot_cellchat(
  mode = "single_pathway",
  slot = "net",
  pathway_show = c("AGRN"),
  lr_show = "WNT4_FZD3_LRP6",
  ident1 = c(2),
  ident2 = c(3)
)
}
\arguments{
\item{db}{string}
}
\value{
json
}
\description{
Run CellChat
}
