% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{plot_enrichr_dot}
\alias{plot_enrichr_dot}
\title{Generate enrichment dot plot}
\usage{
plot_enrichr_dot(df = mtcars, isPvalLog = "true2")
}
\value{

}
\description{
Generate enrichment dot plot
}
