% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Generate GSEA plot}
\usage{
plot_gsea(req, term)
}
\arguments{
\item{req}{request payload}

\item{term}{string}
}
\value{

}
\description{
Generate GSEA plot
}
