% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_deepmaps}
\alias{read_deepmaps}
\title{Read file switcher}
\usage{
read_deepmaps(
  mimetype = "application/vnd.ms-excel",
  path = "c57e41d078ce9810717de392b4451605"
)
}
\arguments{
\item{mimetype}{string}

\item{path}{string}

\item{originalname}{string}
}
\value{
dataframe/matrix/sparse matrix
}
\description{
Read file switcher
}
\references{
https://rviews.rstudio.com/2019/08/13/plumber-logging/
}
