% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_deepmaps_h5}
\alias{read_deepmaps_h5}
\title{Read 10x hdf5 format}
\usage{
read_deepmaps_h5(path)
}
\arguments{
\item{path}{string}
}
\value{
dataframe/matrix/sparse matrix
}
\description{
Read 10x hdf5 format
}
\references{
https://rviews.rstudio.com/2019/08/13/plumber-logging/
}
