% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_deepmaps_text}
\alias{read_deepmaps_text}
\title{Read csv,tsv,txt format}
\usage{
read_deepmaps_text(path)
}
\arguments{
\item{path}{string}
}
\value{
dataframe/matrix/sparse matrix
}
\description{
Read csv,tsv,txt format
}
\references{
https://rviews.rstudio.com/2019/08/13/plumber-logging/
}
