% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_service.R
\name{regulon_enrichment}
\alias{regulon_enrichment}
\title{Run GSEA enrichment}
\usage{
regulon_enrichment(req, genes = c("CD74", "CD7"), database = "C2")
}
\arguments{
\item{req}{request payload}

\item{genes}{array}

\item{database}{string}
}
\value{

}
\description{
Run GSEA enrichment
}
