% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{rename_idents}
\alias{rename_idents}
\title{Rename clusters/idents}
\usage{
rename_idents(req, old_name = 1, new_name = 1)
}
\arguments{
\item{req}{request payload}

\item{old_name}{array}

\item{new_name}{array}
}
\value{
array levels of renamed new idents
}
\description{
Rename clusters/idents
}
