% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_service.R
\name{run_cellchat}
\alias{run_cellchat}
\title{Run CellChat}
\usage{
run_cellchat(job = "single_rna_example", active_idents = "hgt_cluster")
}
\arguments{
\item{active_idents}{string}
}
\value{
json
}
\description{
Run CellChat
}
