% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{select_category}
\alias{select_category}
\title{Set or select category to assign labels}
\usage{
select_category(req, categoryName = "other")
}
\arguments{
\item{req}{request payload}

\item{categoryName}{string}
}
\value{
active category name and levels, all available categories
}
\description{
Set or select category to assign labels
}
