% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{select_cells}
\alias{select_cells}
\title{Select cells based on criteria}
\usage{
select_cells(req, newLevelName = "ct1", filterPayload)
}
\arguments{
\item{req}{request payload}

\item{newLevelName}{array}

\item{filterPayload}{array}
}
\value{
filtered cell labels, obj will store current cells
}
\description{
Select cells based on criteria
}
