% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket_service.R
\name{send_progress}
\alias{send_progress}
\title{Send job progress to server}
\usage{
send_progress(message = "message from R1")
}
\arguments{
\item{message}{string}
}
\description{
Send job progress to server
}
