% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_service.R
\name{set_assay}
\alias{set_assay}
\title{Set Seurat Assay by name}
\usage{
set_assay(req, name = "RNA")
}
\arguments{
\item{req}{request payload}

\item{name}{string idents name}
}
\value{
array levels of new idents
}
\description{
Set Seurat Assay by name
}
