% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_service.R
\name{set_embedding}
\alias{set_embedding}
\title{Set embedding/dimension reduction by name}
\usage{
set_embedding(req, name = "pca")
}
\arguments{
\item{req}{request payload}

\item{name}{string idents name}
}
\value{
array levels of the active embedding
}
\description{
Set embedding/dimension reduction by name
}
