% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_service.R
\name{set_idents}
\alias{set_idents}
\title{Set Seurat Idents by name}
\usage{
set_idents(req, name = "orig.ident")
}
\arguments{
\item{req}{request payload}

\item{name}{string idents name}
}
\value{
array levels of new idents
}
\description{
Set Seurat Idents by name
}
