% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_if_not_null}
\alias{set_if_not_null}
\title{Set a default value if an object is NOT null}
\usage{
set_if_not_null(lhs, rhs)
}
\arguments{
\item{lhs}{An object to set if it's NOT null}

\item{rhs}{The value to provide if x is NOT null}
}
\value{
lhs if lhs is null, else rhs
}
\description{
Set a default value if an object is NOT null
}
\references{
https://adv-r.hadley.nz/functions.html#missing-arguments
}
\author{
Hadley Wickham
}
