% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_service.R
\name{set_obj}
\alias{set_obj}
\title{Set subset obj or full_obj}
\usage{
set_obj(req, type = "full")
}
\arguments{
\item{req}{request payload}

\item{type}{string idents name}
}
\value{
array levels of new idents
}
\description{
Set subset obj or full_obj
}
