% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_entry.R
\name{start_debug_server}
\alias{start_debug_server}
\title{Start debug server}
\usage{
start_debug_server(port = 8000, host = "127.0.0.1")
}
\arguments{
\item{port}{default: 8000}

\item{host}{default: "127.0.0.1"}
}
\description{
Start debug server
}
\examples{
\dontrun{
start_debug_server(port = 8000)
}
}
