% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{static_heatmap}
\alias{static_heatmap}
\title{Complex heatmap}
\usage{
static_heatmap(req, genes = c("CD74", "CD7"), meta = "cell_type", color = NULL)
}
\arguments{
\item{req}{request payload}

\item{genes}{vector of string}

\item{meta}{metadata to annotate in columns}

\item{color}{color palette}
}
\value{
static image
}
\description{
Complex heatmap
}
