% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{subset_cells}
\alias{subset_cells}
\title{Select cells based on criteria}
\usage{
subset_cells(req, selectionPayload)
}
\arguments{
\item{req}{request payload}

\item{selectionPayload}{array}
}
\value{
select usbset cell labels, obj will store current cells
}
\description{
Select cells based on criteria
}
