% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{umap_plot}
\alias{umap_plot}
\title{Plot static UMAP}
\usage{
umap_plot(req, categoryName = "hgt_cluster")
}
\arguments{
\item{req}{request payload}

\item{categoryName}{string}
}
\value{
static image
}
\description{
Plot static UMAP
}
