% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{UploadInfo-class}
\alias{UploadInfo-class}
\title{Create UploadInfo object}
\description{
Create UploadInfo object
}
\section{Slots}{

\describe{
\item{\code{type}}{character}

\item{\code{expr_filename}}{character}

\item{\code{label_filename}}{character}
}}

