% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_service.R
\name{violin_gene_plot}
\alias{violin_gene_plot}
\title{Plot violin gene UMAP}
\usage{
violin_gene_plot(req, gene = "Gad1", split = "sex")
}
\arguments{
\item{req}{request payload}

\item{gene}{string}

\item{split}{string}
}
\value{
static image
}
\description{
Plot violin gene UMAP
}
