% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yan_2013}
\alias{yan_2013}
\title{Data of yan_2013}
\format{
A list with expression matrix and metadata:

$meta:
\describe{
\item{$expr}{a data frame with 20214 genes (rows) amd 90 cells (columns)}
\item{$meta$Cluster}{fct cell type}
}
}
\source{
\url{https://www.nature.com/articles/nsmb.2660}
}
\usage{
yan_2013
}
\description{
Yan 2013: Human embryo, 7 cell types, 90 cells
}
\keyword{datasets}
