% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeisel_2015}
\alias{zeisel_2015}
\title{Data of zeisel_2015}
\format{
A list with expr and labels:
$expr: a data frame with 19972 genes (rows) amd 3005 cells (columns)
$meta:
#'\describe{
\item{Label}{fct cell type}
\item{Sex}{fct random generated sex info}
\item{Sample}{fct random generated sample info}
}
}
\source{
\url{https://science.sciencemag.org/content/347/6226/1138.abstract}
}
\usage{
zeisel_2015
}
\description{
Zeisel 2015: Mouse Brain, 7 cell types, 3005 cells
}
\keyword{datasets}
