% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_service.R
\name{calc_deg}
\alias{calc_deg}
\title{Run Seurat clustering}
\usage{
calc_deg(
  req,
  ident1 = c(0, 1),
  ident2 = c(2, 3),
  min_pct = 0.2,
  min_lfc = 0.5,
  assay = "RNA",
  pvalue = 0.05,
  direction = "all"
)
}
\arguments{
\item{req}{request payload}

\item{ident1}{string}

\item{ident2}{string}

\item{min_pct}{string}

\item{min_lfc}{string}

\item{assay}{string}

\item{pvalue}{string}

\item{direction}{string}
}
\value{

}
\description{
Run Seurat clustering
}
