% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{cluster_multiome}
\alias{cluster_multiome}
\title{Run multiome clustering}
\usage{
cluster_multiome(
  req,
  jobid = "example",
  method = "Velocity weighted method",
  nPCs = "20",
  resolution = "0.5",
  neighbor = "20"
)
}
\arguments{
\item{req}{request payload}

\item{nPCs}{string}

\item{resolution}{string}

\item{neighbor}{string}
}
\value{

}
\description{
Run multiome clustering
}
