% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_service.R
\name{example_regulon_network}
\alias{example_regulon_network}
\title{Run Seurat clustering}
\usage{
example_regulon_network(cluster = "sda")
}
\arguments{
\item{cluster}{string}
}
\value{
json
}
\description{
Run Seurat clustering
}
