% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_service.R
\name{load_multiome}
\alias{load_multiome}
\title{Load matched scRNA-seq and scATAC-seq data to Seurat / Signac}
\usage{
load_multiome(
  req,
  idx = 0,
  filename,
  min_cells = 1,
  min_genes = 200,
  nVariableFeatures = 3000,
  percentMt = 5,
  removeRibosome = FALSE
)
}
\arguments{
\item{req}{request payload}

\item{idx}{sample index}

\item{filename}{string}

\item{min_cells}{number}

\item{min_genes}{number}

\item{nVariableFeatures}{number}

\item{percentMt}{number}

\item{removeRibosome}{boolean}
}
\value{
list of basic QC metrics
}
\description{
Load matched scRNA-seq and scATAC-seq data to Seurat / Signac
}
