% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print_wd}
\alias{print_wd}
\title{Print working directory and a message}
\usage{
print_wd(msg = "Hello")
}
\arguments{
\item{msg}{a character message}
}
\value{
a character message
}
\description{
Print working directory and a message
}
