% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_service.R
\name{select_cells}
\alias{select_cells}
\title{Select cells based on criteria}
\usage{
select_cells(req, newLevelName = "ct1", filterPayload)
}
\arguments{
\item{req}{request payload}

\item{newLevelName}{array}

\item{filterPayload}{array}
}
\value{
filtered cell labels, obj will store current cells
}
\description{
Select cells based on criteria
}
\examples{
\dontrun{
filterPayload <- data.frame(
  type = c("gene", "cluster"),
  name = c("Gad1", NA),
  direction = c(">", "in"),
  thres = c("1", NA),
  category = c(NULL, "cell_type"),
  level = c(NA, "1_oligodendrocytes")
)
newLevelName <- "label1"

this_filter <- list(
  type = "cluster",
  direction = "in",
  category = "cell_type",
  level = "1_oligodendrocytes"
)
}

}
