% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_entry.R
\name{start_server}
\alias{start_server}
\title{Start IRIS3 R api server}
\usage{
start_server(port = 8000, host = "127.0.0.1")
}
\arguments{
\item{port}{The port on which to serve the output viewer.  If omitted, a
random port between 8000 and 9999 is chosen.}

\item{host}{The host used to serve the output viewer. If omitted, "127.0.0.1"
is used.}
}
\description{
Start IRIS3 R api server
}
\examples{
\dontrun{
start_server(port = 8000)
}
}
