% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_entry.R
\name{start_server}
\alias{start_server}
\title{Start IRIS3 R api server}
\usage{
start_server(port = 8000, host = "127.0.0.1")
}
\arguments{
\item{port}{default: 8000}

\item{host}{default: "127.0.0.1"
is used.}
}
\description{
Start IRIS3 R api server
}
\examples{
\dontrun{
start_server(port = 8000)
}
}
