% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{UploadInfo-class}
\alias{UploadInfo-class}
\title{Create UploadInfo object}
\description{
Create UploadInfo object
}
\section{Slots}{

\describe{
\item{\code{type}}{ANY.}

\item{\code{expr_filename}}{ANY.}

\item{\code{label_filename}}{ANY.}
}}

